Fix bugs in tests and save_chartdata
Updates a number of tests that were failing because of deprecation of the 'blog' chart style, the introduce of cairo as the default device, and because of changed behaviour of grattan_save_pptx.

Also updates grattan_save and save_chartdata so it properly handles cases where grattan_save is used to output chart data and select_data is set to false.