Add handling of null patchwork objects for save_chartdata
Previously, save_chartdata would throw an error if no chartdata was found in the object. This handling worked ok for standard ggplot charts, but doesn't work for patchwork charts when plot_spacer() is used to create extra space in between separate charts. This change better handles this situation and creates clearer warnings.