Change warnings for save_chartdata
Previously save_chartdata threw a warning if select_data was TRUE and a column that was used in a layer wasn't found in the chart data. However, this occurs often when layers have their own data that isn't in the underlying ggplot dataframe. Instead, the function now checks only in layers that don't rely on their own dataframe.